+++ src/services/Assets.php
@@ -516,9 +516,10 @@ class Assets extends Component
      *
      * @param VolumeFolder $parentFolder
      * @param string $orderBy
+     * @param bool $withParent Whether the parent folder should be included in the results
      * @return array
      */
-    public function getAllDescendantFolders(VolumeFolder $parentFolder, string $orderBy = 'path'): array
+    public function getAllDescendantFolders(VolumeFolder $parentFolder, string $orderBy = 'path', bool $withParent = true): array
     {
         /** @var Query $query */
         $query = $this->_createFolderQuery()
@@ -531,6 +532,10 @@ class Assets extends Component
 
         if ($orderBy) {
             $query->orderBy($orderBy);
+        }
+
+        if (!$withParent) {
+            $query->andWhere(['not', ['id' => $parentFolder->id]]);
         }
 
         $results = $query->all();